[UPDATE] move HomePage methods
