+++ src/HomePage/HomePageSettings.php
@@ -8,8 +8,7 @@ use Epesi\Core\HomePage\Database\Models\HomePage;
 use Epesi\Core\Layout\Seeds\ActionBar;
 use Epesi\Core\System\Seeds\Form;
 use Spatie\Permission\Models\Role;
-use atk4\data\Persistence;
-use atk4\data\Persistence_Array;
+use Epesi\Core\HomePage\Integration\Joints\HomePageJoint;
 
 class HomePageSettings extends ModuleView
 {
@@ -19,9 +18,16 @@ class HomePageSettings extends ModuleView
 	protected $grid;
 	protected $form;
 	
+	/**
+	 * Fallback path in case no home page set for the user
+	 *
+	 * @var string
+	 */
+	protected static $defaultPath = 'view/user.settings';
+	
 	public static function access()
 	{
-		return Auth::user()->can('modify system settings') && HomePageCommon::getAvailableHomePages();
+		return Auth::user()->can('modify system settings') && self::getAvailableHomePages();
 	}
 	
 	public function body()
@@ -41,7 +47,7 @@ class HomePageSettings extends ModuleView
 		
 		$this->grid->setModel($this->getModel());
 		
-		$availablePages = HomePageCommon::getAvailableHomePages();
+		$availablePages = self::getAvailableHomePages();
 
 		$this->grid->addDecorator('path', ['Multiformat', function($row, $column) use ($availablePages) {
 			return [['Template', $availablePages[$row[$column]]?? '[' . __('missing: :path', ['path' => $row[$column]]) . ']']];
@@ -67,7 +73,7 @@ class HomePageSettings extends ModuleView
 	
 	public function getModel()
 	{
-		$availablePages = HomePageCommon::getAvailableHomePages();
+		$availablePages = self::getAvailableHomePages();
 		$availableRoles = Role::get()->pluck('name')->all();
 
 		$rows = [];
@@ -122,5 +128,48 @@ class HomePageSettings extends ModuleView
 		}
 
 		return $this->form;
+	}
+	
+	/**
+	 * Collect all home pages from module joints
+	 *
+	 * @return array
+	 */
+	public static function getAvailableHomePages()
+	{
+		static $cache;
+		
+		if (! isset($cache)) {
+			$cache = [];
+			foreach (HomePageJoint::collect() as $joint) {
+				$cache[$joint->link()] = $joint->caption();
+			}
+		}
+		
+		return $cache;
+	}
+	
+	/**
+	 * Get the current user home page
+	 *
+	 * @return HomePage
+	 */
+	public static function getUserHomePage()
+	{
+		if (! $user = Auth::user()) return;
+		
+		return HomePage::whereIn('role', $user->roles()->pluck('name'))->orderBy('priority')->first();
+	}
+	
+	/**
+	 * Get the current user home page path
+	 *
+	 * @return HomePage
+	 */
+	public static function getUserHomePagePath()
+	{
+		$homepage = self::getUserHomePage();
+		
+		return $homepage? $homepage->path: self::$defaultPath;
 	}
 }
